Update 04/10/2021: Overhauled event observing system, removed unnecessary scripts and prefabs, bullet ricochet noises, damage markers and a melee attack. Lots of work done, hooray!
Overhauled EventHandler/EventObserver system and replaced it with an EventJunction system. This new system is much more efficient as it removes the EventObserver middleman (EventHandler still exists with some code as a proof of concept, in case I need to change some things), and objects that need to listen for events can subscribe directly to the EventJunction. As EventJunction is also a static class, no GameObject for it needs to exist in the scene. The new system is also greatly streamlined by removing a lot of unnecessary checks and repeated code, which could have been condensed into multi-use properties and functions. That being said, I need to fix two problems – there is currently no proper system implemented for removing functions that have been added when the component running said function is disabled or destroyed, and there is no system for refreshing the events when a scene transition occurs. Fortunately, I have various ideas for how to fix these things, which I will implement at a later date.

Added the hit marker function back into the game. Previously disabled due to running on awkward and inefficient code using GetComponent to find a PlayerHandler, it now runs in a much more modular manner by listening for entities being damaged using the EventJunction system. It is also slightly tweaked – it still has two marker variants, but instead of playing the elaborate one when performing a critical hit, it is performed when a target is killed.

Added impact noises for when a bullet hits a surface. Since I did not want to increase memory by having an AudioSource on every bullet, I took my RandomSoundPlayer and implemented a function to utilise AudioSource.PlayClipAtPoint with the position of a specified transform, allowing me to implement said function in a UnityEvent and sidestep its GUI restrictions (as AudioSource.PlayClipAtPoint has more than one variable, and some of which do not appear in the UnityEvent window).

Tweaked the Dodge state on the placeholder soldier enemy, so they move at a more reasonable speed and select positions further away from their current one. I briefly tested this and it caused some extra problems I will need to solve, but these were stats I needed to change, so I will have to fix them.

Removed the Prefab variants for projectiles, and figured out a smarter and more elegant way to include gun projectile stats in an easy to alter manner. I have previously tried to implement mechanics such as ScriptableObjects and/or prefab variants in order to allow editing of projectile stats without making a new prefab for every projectile attack, but every solution I could think of still resulted in extra objects that had to be navigated to in order to be edited. I then figured out the simple but elegant solution of including a disabled copy of the original bullet prefab on the weapon itself, disabled and with the appropriate stats assigned. My gun code would then spawn and enable this instance, rather than a prefab in the assets folder. This is a superior system because the bullet prefab is in the same place as the other weapon stat objects, and is therefore easier and more intuitive to access and edit.

Removed a large number of redundant old scripts, mostly older, inferior revisions of AI mechanics.
Removed old knockback and explosion functions from the Damage class. There is now only one function remaining, and once it is rendered redundant I will be able to delete the entire class.

Overhauled the melee attack function again. The current version uses a reworked version of my new explosion code, but with an additional angle check to only damage enemies that are inside a specific angle. I briefly tested the attack with the player character and it feels very intuitive (even though there is no animation), with the player able to easily ‘swing’ at enemies and reliably deal damage to them without requiring precision, since a melee attack in a first-person shooter is generally intended as a backup if the player is presently unable to accurately shoot a close-up enemy.

Fixed a bug where the player could attack enemy corpses and produce damage markers as if they were meaningfully depleted health. I ensured this would only happen if a bool in the health script was enabled to allow postmortem damage (in case I decide to have something that can be destroyed further after death).

Performed a small amount of work on a function for running different events based on the substance an object is made of, gathered using a RaycastHit. I ran into too many quandaries and decided to temporarily pause work on it, in order to not waste too much time and implement other, more important mechanics.