Update 05/10/2021: Volume adjustment, experimenting with recoil animations, removing a useless class, better names for functions
Reduced volume of player footsteps
Changed EventJunction.RefreshWithFunction’s name to Subscribe, to be shorter and easier to type.
Experimented with function for code-based recoil animations for guns, but it didn’t work very well. The code just lerps the gun model to a position and back using an AnimationCurve, but despite using lerping the animation was still very choppy. While I have an idea to implement better cosmetic recoil, I will implement it at a later date.
Removed the EventHandler class, and copied the small section of valuable code in it to the EventJunction class, commented out to prevent it from interfering with anything.