Update 18/10/2021 – 20/10/2021: Bug fixes and backends with objectives and event junctions. Also working on modelling a new level environment.
Added a bit of code to ensure the AI’s Dodge function would be properly removed from the EventJunction when the AI script is disabled. This is necessary to ensure that these functions do not remain in the EventJunction when the AI is killed or during a scene transition.
Fixed a bug where the player’s health meter would not update visually when replenished. However, the process to do so has created a new bug wherein a damage marker will play whenever the player’s health is restored.
Started experimenting with modelling a new game environment, both to practice my 3D modelling skills and to have a prettier environment for demonstration purposes than the mess of primitive objects that I currently have.
Experimented with cosmetic gun recoil using a rigidbody and Joint component, which did not work.
Reworked the player objective system to make it easier to insert data. For example, the KillGroupObjective, that completes when the player kills a specified set of enemies in the level, is now updated to automatically obtain the enemy values using GetComponent. This allows me or any other developer working on the game to simply drag the enemies into the objective as children, without having to manually add them to the array.