Update 21/10/2021: Testing a placeholder level, actual weapon models, picking up weapons, balancing, weapon selector and interaction reworking
Created a new level for testing, using the map model I was working on previously. Although the level is very bare and doesn’t fully work from a gameplay perspective, I am very pleased with the job I did of prototyping a somewhat immersive looking environment, despite my limited 3D modelling experience.
Started implementing new versions of the existing weapons with actual models. Still no animations, but I also plan on implementing proper animation controllers to make the animations for firing, reloading, drawing and holstering.
Started implementing functionality for picking up and dropping guns, since there needs to be a method for acquiring new weapons rather than just having them at the start.
Reduced handgun and rifle spread, for better balancing.
Reworked much of the code in the radial menu script, since it was confusingly written and causing lots of bugs.
Slightly reworked the Interactable object in order to run more complex commands in its UnityEvent, by allowing it to use the player who interacted with it as a parameter. I did this by creating a custom class deriving from UnityEvent, with an additional parameter for a PlayerHandler. My original intention was to just reference a ‘UnityEvent<PlayerHandler>’ by itself, but this resulted in the field not showing in the Inspector. With this method, I could add code to my guns to have the player pick them up upon interaction, without having to make an inherited version of the Interactable class for such a purpose. Thanks to this change, I may also be able to modify my other Interactable objects to not need inherited classes.